@@ -12,15 +12,14 @@
 
 package net.ftlines.wicket.fullcalendar.callback;
 
-import net.ftlines.wicket.fullcalendar.CalendarResponse;
-import net.ftlines.wicket.fullcalendar.ViewType;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.request.Request;
-import org.joda.time.DateMidnight;
-import org.joda.time.DateTimeZone;
-import org.joda.time.format.DateTimeFormatter;
-import org.joda.time.format.ISODateTimeFormat;
+
+import net.ftlines.wicket.fullcalendar.CalendarResponse;
+import net.ftlines.wicket.fullcalendar.ViewType;
 
 /**
  * A base callback that passes back calendar's starting date
@@ -31,10 +30,8 @@ import org.joda.time.format.ISODateTimeFormat;
 public abstract class ViewDisplayCallback extends AbstractAjaxCallback implements CallbackWithHandler {
 	@Override
 	protected String configureCallbackScript(String script, String urlTail) {
-		return script
-			.replace(
-				urlTail,
-				"&view=\"+v.name+\"&start=\"+fullCalendarExtIsoDate(v.start)+\"&end=\"+fullCalendarExtIsoDate(v.end)+\"&visibleStart=\"+fullCalendarExtIsoDate(v.visStart)+\"&visibleEnd=\"+fullCalendarExtIsoDate(v.visEnd)+\"");
+		return script.replace(urlTail,
+			"&view=\"+v.name+\"&start=\"+fullCalendarExtIsoDate(v.start)+\"&end=\"+fullCalendarExtIsoDate(v.end)+\"&visibleStart=\"+fullCalendarExtIsoDate(v.visStart)+\"&visibleEnd=\"+fullCalendarExtIsoDate(v.visEnd)+\"");
 	}
 
 	@Override
@@ -46,15 +43,14 @@ public abstract class ViewDisplayCallback extends AbstractAjaxCallback implement
 	protected void respond(AjaxRequestTarget target) {
 		Request r = target.getPage().getRequest();
 		ViewType type = ViewType.forCode(r.getRequestParameters().getParameterValue("view").toString());
-		DateTimeFormatter fmt = ISODateTimeFormat.dateTimeParser().withZone(DateTimeZone.UTC);
-		DateMidnight start = fmt.parseDateTime(r.getRequestParameters().getParameterValue("start").toString())
-			.toDateMidnight();
-		DateMidnight end = fmt.parseDateTime(r.getRequestParameters().getParameterValue("end").toString())
-			.toDateMidnight();
-		DateMidnight visibleStart = fmt.parseDateTime(
-			r.getRequestParameters().getParameterValue("visibleStart").toString()).toDateMidnight();
-		DateMidnight visibleEnd = fmt
-			.parseDateTime(r.getRequestParameters().getParameterValue("visibleEnd").toString()).toDateMidnight();
+		DateTimeFormatter fmt = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss'Z'");
+		// DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+		LocalDateTime start = LocalDateTime.parse(r.getRequestParameters().getParameterValue("start").toString(), fmt);
+		LocalDateTime end = LocalDateTime.parse(r.getRequestParameters().getParameterValue("end").toString(), fmt);
+		LocalDateTime visibleStart = LocalDateTime
+			.parse(r.getRequestParameters().getParameterValue("visibleStart").toString(), fmt);
+		LocalDateTime visibleEnd = LocalDateTime
+			.parse(r.getRequestParameters().getParameterValue("visibleEnd").toString(), fmt);
 		View view = new View(type, start, end, visibleStart, visibleEnd);
 		CalendarResponse response = new CalendarResponse(getCalendar(), target);
 		onViewDisplayed(view, response);
